Updates in lyap_r in measures.py
We added an extra argument called fit_offset (int) which allows for neglecting the first fit_offset steps when fitting div_traj.
In addition, we changed for i in range(n) to for i in range(1,n) when determining the lag. Otherwise, the last criterion of the if clause always is fulfilled and thus leads to a lag = 0.